requirements.txt
openai-whisper
bs4
tqdm
requests
lxml
lunr
